exercise 4.21

+++ part4/exercises/blogposts/controllers/blogs.js
@@ -53,7 +53,21 @@ blogsRouter.post('/', async (request, response) => {
 
 blogsRouter.delete('/:id', async (request, response, next) => {
   try {
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!decodedToken.id) {
+      return response.status(401).json({ error: 'token invalid' })
+    }
+
+    const user = await User.findById(decodedToken.id)
+
+    const blog = await Blog.findById(request.params.id);
+
+    if (blog?.user?.toString?.() !== user.id.toString() && !blog?.user?.id?.toString?.() !== user.id.toString()) { 
+      return response.status(401).json({ error: 'user not authorized to delete this blog post' });
+    }
+
     await Blog.findByIdAndDelete(request.params.id);
+
     response.status(204).end();
   } catch (error) {
     next(error);
